requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
diagrams==0.17.0
Django==3.1.1
django-crispy-forms==1.9.2
djangorestframework==3.11.1
graphviz==0.13.2
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
Pillow==7.2.0
pymongo==3.11.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
